+++ src/modules/gear/components/ColorAutocomplete.vue
@@ -14,9 +14,15 @@ const modelValue = defineModel<string>('value', { default: '' })
     :create-label="$t('gear.comboBox.add')"
     class="w-full"
   >
+    <template #value>
+      <div class="flex items-center gap-2">
+        <div class="size-4 rounded-full" :style="{ backgroundColor: modelValue as string }" />
+        <span>{{ modelValue }}</span>
+      </div>
+    </template>
     <template #option-content="{ option }">
       <div class="flex items-center gap-2">
-        <div class="w-4 h-4 rounded-full" :style="{ backgroundColor: option.data as string }" />
+        <div class="size-4 rounded-full" :style="{ backgroundColor: option.data as string }" />
         <span>{{ option.label }}</span>
       </div>
     </template>